refactor(core): separate system record identity codecs
